configure.ac: do not override user variables
CXXFLAGS should not be messed with. They're for the user to set.
Use AM_CXXFLAGS instead. See:
http://www.gnu.org/software/automake/manual/html_node/Flag-Variables-Ordering.html

But use it only to set "-Wall". Let's be nice to downstreams and don't
set "-Werror" by default. See for example:
https://lists.fedoraproject.org/pipermail/devel/2011-March/149937.html

I don't see any explanation or justification for setting the
inline-unit-growth gcc parameter, so let's drop that too.

I suspect checking here for $TARGET = FREEBSD was bogus to being with.
Probably the actual reason for differentiating was gcc vs. clang, not
FreeBSD vs. whatever.

Defining __LINUX__ makes no sense.
Defining VERSION is redundant. It's already in config.h.