issue: 1801987 Add sanity check for the sockets list
Additionally to flat array of sockets file descriptors, sockperf links
the sockets into a cyclic linked list. Wrapping the list is done as
linking the maximum socket number with the minimum one.

The algorithms works while OS allocates file descriptors in an increasing
order. If fd was allocated out of the order the algorithm would possibly
produce an invalid list, for example:

 4 -> 3 -> 6 -+  5
      ^       |
      +-------+

Such a scenario can happen if a thread opens and closes fds concurrently
with sockperf. This is achieved through LD_PRELOAD.

Add sanity check for the cyclic list and report an error instead of
processing an invalid list.

Signed-off-by: Dmytro Podgornyi <dmytrop@mellanox.com>