issue: 1801987 Fix segfault with an empty feed file
Socket objects are linked into a cyclic linked list. The operation
of looping the list leads to a NULL pointer dereference if the list
is empty.

Exit with error when user provides an empty feeedfile. File which
contains only comments and empty lines is considered empty as well.

Also move the block which checks for an I/O error out of the reading loop.
The block was unreachable, because (!res) is condition when the loop
stops.

Signed-off-by: Dmytro Podgornyi <dmytrop@mellanox.com>