issue: 3002911 Fix error when compiling OpenSSL 3.0.0
OpenSSL development(alpha) versions for 3.0.0 release deprecated lot's
of previously supported APIs before adding the new ones.
Additionally, there is no simple way to distinguish between 3.0.0 alpha
and release versions since the OPENSSL_VERSION_NUMBER is of the same
value. Thus, the only reasonable solutions seems to be the one where we
migrate to the OpenSSL 3 APIs for versions that after the 3.0.0.
Potential "API deprecated" warnings will be suppressed for previous
versions.

Signed-off-by: Alex Briskin <abriskin@nvidia.com>