issue: 3014055 Replace select(2) in the prepare phase
If connect(2) returns a recoverable error, we try to poll for an event
and check error on the socket with getsockopt(SO_ERROR). Originally this
was done using select(2) which doesn't support file descriptor value
greater than FD_SETSIZE which is 1024.

A socket fd greater or equal to 1024 makes select(2) fail and this fail
isn't handled. In turn, this makes getsockopt(SO_ERROR) return
EINPROGRESS since connection isn't established yet and sockperf
terminates with the message "Can`t connect socket".

As solution for this issue and limitation, replace select(2) with
poll(2). Similarly to original behavior, don't terminate on polling
errors. Terminate only if getsockopt() reports an error.

For Windows keep old select(2) implementation, because it has different
poll(2) API and needs porting.

Signed-off-by: Dmytro Podgornyi <dmytrop@nvidia.com>