[CI] Adjust timeouts for tests
Due to the significant load on the Jenkins agent on the baremetal host,
the timeouts for tests and the pipeline are insufficient to process all
test cases. Here we temporarily extend the timeouts to let CI runs, but
later we're going to move some tests to containers: HPCINFRA-3252 to
improve timings and to decrease the load to baremetal host.

Signed-off-by: Viacheslav Login <vlogin@nvidia.com>